+++ src/components/PaginationButton/styles.ts
@@ -3,12 +3,12 @@ import { marginPaddingReset } from "@/styles";
 import { PaginationButtonProps, StyledPaginationButton } from "./types";
 
 export const StyledWrapper: IStyledComponent<any, any> = styled.div<
-  Pick<PaginationButtonProps, "intent">
+  { $intent: PaginationButtonProps["intent"] }
 >`
   padding: 4px 11px 0;
   width: 100%;
-  border-top: ${({ intent, theme }) =>
-    intent === "next" ? `1px solid ${theme.colors.border}` : "none"};
+  border-top: ${({ $intent, theme }) =>
+    $intent === "next" ? `1px solid ${theme.colors.border}` : "none"};
 `;
 
 export const StyledButton: IStyledComponent<any, any> = styled.button<StyledPaginationButton>`
@@ -25,8 +25,8 @@ export const StyledButton: IStyledComponent<any, any> = styled.button<StyledPagi
   line-height: 150%;
   letter-spacing: 1px;
   cursor: pointer;
-  opacity: ${({ isVisible }) => (isVisible ? "1" : "0")};
-  pointer-events: ${({ isVisible }) => (isVisible ? "auto" : "none")};
+  opacity: ${({ $isVisible }) => ($isVisible ? "1" : "0")};
+  pointer-events: ${({ $isVisible }) => ($isVisible ? "auto" : "none")};
   &:hover {
     transition: 0.5s ease;
     background-color: ${({ theme }) => theme.colors.hover};