+++ src/index.ts
@@ -1,6 +1,15 @@
 import { Scheduler } from "./components";
 import "./styles.css";
 export type { SchedulerProps } from "./components/Scheduler/types";
-export type { SchedulerData, SchedulerProjectData, ZoomLevel, Config } from "./types/global";
+export type {
+  SchedulerData,
+  SchedulerProjectData,
+  ZoomLevel,
+  Config,
+  CellClickData,
+  CellRangeSelectData,
+  SelectedRange,
+  TileMoveData
+} from "./types/global";
 
 export { Scheduler };